backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
chromadb==0.4.24
google-generativeai==0.4.0
pydantic==2.6.3
python-multipart==0.0.9
pandas==2.2.1
sentence-transformers==2.5.1
mlflow==2.11.1
boto3==1.34.69
scikit-learn==1.4.1.post1
PyJWT==2.8.0
passlib[recommended]==1.7.4
bcrypt==4.1.2
